pyproject.toml
[tool.poetry]
name = "exaflow"
version = "1.0.0"
description = "Core engine for the Medical Informatics Platform"
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "exaflow" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hypercorn = "~0.17"
Quart = "~0.20"
asgiref = "~3.9"
pandera = "~0.25"
duckdb = "~1.1"
duckdb-engine = "~0.11"
click = "~8.1"
toml = "~0.10"
SQLAlchemy = "~2.0"
astor = "~0.8"
numpy = "2.2.6"
scipy = "1.15.3"
pandas = "2.3.1"
scikit-learn = "1.7.0"
dnspython = "~2.7"
pydantic = "~1.10"
envtoml = "~0.1"
requests = "~2.32"
patsy = "~0.5.3"
psutil = "~5.9"
xgboost-cpu = "^2.1.1"
pyarrow = "^22.0.0"
grpcio = "1.73.1"
protobuf = "4.25.8"
flwr = "^1.23.0"

[tool.poetry.dev-dependencies]
pytest = "~8.4"
pytest-xdist = "~3.8"
pytest-asyncio="~1.1"
pytest-mock="~3.14"
pytest-assume="~2.4"
ipython = "~8.37"
pdbpp = "~0.11"
invoke = "~2.2"
black = "25.1.0" # Must be upgraded together with github actions and pre-commit configuration
isort = "6.0.1" # Must be upgraded together with github actions and pre-commit configuration
pre-commit = "~4.2"
termcolor = "~3.1"
coverage = {extras = ["toml"], version = "^7.9.2"}
pytest-cov = "~6.2"
pylint = "~3.3"
rope = "~1.14"
devtools="~0.12"
docker = "~7.1"
tqdm = "~4.67"
statsmodels = "0.14.5"
psutil = "~5.9"
pytest-subtests = "~0.14"
freezegun = "~1.5"
hypothesis = "~6.135"
pytest-rerunfailures = "~15.1"

[tool.poetry.group.aggregation-server.dependencies]
grpcio = "~1.73"
tomli = "~2.2"
grpcio-health-checking = "~1.62"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow, mostly integration tests (deselect with '-m \"not slow\"')",
    "very_slow: marks tests as very_slow, integration tests with temporary container dependencies (deselect with '-m \"not very_slow\"')",
    "smpc: marks the tests validating smpc functionality (deselect with '-m \"not smpc\"')",
    "smpc_cluster: marks the tests that need smpc cluster deployment (deselect with '-m \"not smpc_cluster\"')",
]
filterwarnings = ["ignore::DeprecationWarning"]
norecursedirs = ["tests/testcase_generators"]

[tool.coverage.run]
source_pkgs = ['exaflow']
dynamic_context = 'test_function'

[tool.coverage.report]
exclude_lines = ['def __repr__', 'raise NotImplementedError', 'pragma: no cover']
show_missing = true

[tool.coverage.html]
show_contexts = true
